Enable the new mman by default
It's finally time: the new mman can run user-space as far as I can tell, and nothing seems to outright crash. Some features are still missing, and the old VM will stick around while little longer, and kernel code will continue to support the old VM system for a while longer still. But with this, the new mman oficially becomes the new standard!