Redesign exception handling of driver init/fini functions
In the interest of driver initialization & finalization both being able to throw exceptions, as well as perform blocking operations, it stands to reason that CTRL+C'ing such an operation should also allow that operation to be re-started. As such, both kinds of callbacks need to be able to throw exceptions, as well as have them be propagated.