Fix a couple more problems with the new driver system
With this, the kernel is already able to load (some) drivers (in practice: `sctrace' and `procfs'), but right now the next problem is:
>> [mod][ps2] Relocation against unknown symbol "drv_arg_timeout_infull"
Which is caused by the fact that I havn't implemented name argument symbol support, yet. - However, I do plan on implementing support for this feature differently this time around, namely by including direct support for decoding of argument types, as well as default values.