Fix but in xml-escape decoder
When input data ends with a unicode entity that has an optional trailing ';', and that ';' isn't present, the decoder must use a skip-optional-semicollon state that still counts as a zero-shift state. Prior to this commit, that case wasn't handled and the decoder wouldn't have output the escaped character even though it was present.