New libc function `write_printer(3)'
Pretty much just an alias for `writeall(3)', but is declared as and guarantied to be pformatprinter-compatible. In turn, this function forms a wrapper for `writeall(3)' and may be used as a sink for streaming data into a file descriptor.
In turn, `dprintf(3)' has also been adjusted to make use of this function