Fix a linux compatibility issue
When pread() or pwrite() isn't implemented by a given kernel object, then the resulting exception should be translated to `ESPIPE' ("Illegal seek"), because the associated object doesn't support the neccessary seek operation that would be needed to reach the requested file-location.

The linux kernel does the same, and now KOS does, too.