Get rid of the never-used <format-reader.h>header
Functions using this system (which was meant to be similar to <format-printer.h>, only for input rather than output) were never created, and I believe the reason is that input data can't really be serialized efficiently, the way that output data can. Pretty much always, input data already takes a format most suitable for the specific application.
Additionally, the idea of having a callback that provides data to-be processed bit-by-bit implicitly requires that (to prevent dead-locking) no locks are held while such a callback is invoked. And as a result, anything that would consume such data already would need to be designed to accept data being fed in variable-sized chunks. The final nail in the coffin is that such an API can just be written to expose a `ssize_t feed_me_data(void const *data, size_t num_bytes)'-style function, which can be used more efficiently and doesn't necessitate the use of some pseudo-standard callback prototype.