Optimization to userprocmask
Instead of checking each of the pending signal bits individually, have another flags-field with a bit that is set by the kernel when _any_ of the pending signal bits may have been set to `1'.

This way, we don't need to loop through the pending sigset_t every time signal mask restrictions are loosened, but only if it's known that there are actually some signals pending!