Design a system for user-space IPC
It's still in its infancy and nowhere near fully implemented (it doesn't even get compiled into a library, yet), but once this system is implemented, it will allow individual user-space processes to expose public interfaces which can be loaded by arbitrary other processes.
The idea here is that this service system will allow for implementations of things such as window servers and the like.