Implement `service_open(3)' except for the handshake
With this, the client-side is pretty much fully implemented, with the exception of details relating to out-of-band communication (as used for com interrupts), as well as the startup handshake.

The next step will be to implement the server-side, during which it should become clear how best to facilitate these remaining features.