Handle the case of open(2) with S_IFSOCK files
Specs say to return with errno = ENXIO, so in the case of KOS this requires the addition of a dedicated exception parameter.
In conjunction to this, also extend the `E_ILLEGAL_OPERATION' exception class with a general-purpose "reason" parameter, and update all other places that use `E_ILLEGAL_OPERATION' to include their own (unique) reason code, too.