Implement restart select for user RPC programs
Also remove the ability to set flags for restarting system calls that should never be restarted (such as `pause(2)'). There's a point in these being marked as such, and that aside: user RPC programs are already powerful enough to restart such a system call if they really want to, but by adding a flag that would do exactly that, the wrong message is send, as such a flag could never really be used safely!