Remove register order support from libunwind
At the time, I thought I was on to something with this, but as it later turned out: it _really_ doesn't matter in which order you choose to unwind registers. So keeping track of any sort of ordering system here is just a waste of clock cycles.

With this, unwinding will still work just as well as it did before, but should now be a little bit faster!