Document the use of epoll RPCs for detecting HUP
This is what it's all been leading up to:
 - The re-write of the kernel RPC system, which ended up growing to the point of also needing to re-write the user<=>kernel exception handling system (though granted: that one had been in dire need of a re-write already)
 - The extension of the epoll API to allow for RPC delivery as the result of raising-edge conditions in arbitrary files
And with all of this combined (iow: the culmination of 2 weeks work), it's finally possible to provide a super-efficient way of asynchronously monitoring socket files for HUPs (which is needed by libservice). Though obviously: epoll RPCs are much more powerful, and can be used for much more than just something like this!