Fix bug in `ext2_lnk_v_readlink()'
Urgh!... At least I was able to spot that `l = strlen(symname)+1;' in the link source code. BUT WHY NOT CALL IT `s'? `l' I immediatly understand as `length', but that's incorrect; it's the SIZE of the link text (with a trailing NUL that gets written to disk for whatever reason)
But at least the symlink file size is the length _WITHOUT_ a trailing NUL!