Implement a fallback-system for finding the root partition
This one really should be used, as it can significantly degrade performance. - Essentially, this system will try to open all block devices it finds as superblocks before trying to access a file "/os/kernel.bin". If that file exists and appear to match the currently running kernel binary, the located filesystem is then mounted under "/"