Add ABI compat for `extern int errno;'
By doing some magic with IDATA symbols and self-modifying code, KOS's libc is able to re-configure itself when the primary application tries to import a symbol by the name of `errno'.

Note however that doing this, anything that uses errno will no longer be thread-safe (obviously: since there's now only a single, global `errno' variable), and that libc will literally re-write its own assembly in order to facilitate this.

As such, this feature is only available on i386 (but not on x86_64), since it's mainly meant for compatibility with old linux programs.