Refactor libsvga to libsvgadrv
Though the plan was to rename it to `libvga', I felt like that name was too generic (and I bet there's already a library by that name which I may one day want to port to KOS). As such, and given that this library is mainly meant for the implementation of video chipset drivers, this is a much more fitting name.