Add support for KOS-style futexfd objects
These aren't like linux's: these are actually useful since they allow you to specify expressions which are tested during a poll.
Combined with the ability to specifiy more than one expression, these are just as powerful as the lfutexexpr(2) system call, only that they can additionally be used with poll(2) and even epoll(2)!