Fix ata detect code sporatically failing
I'm guessing this is another one related to me updating qemu, but in this case it actually seems to be KOS's fault.

Reading the documentation again, these device-kind codes only get reported after you did ATA_COMMAND_IDENTIFY. As such, only doing the init for special values is undefined and only worked because all of the emulators I've ever used had consistent values after a reset. But now it seems as though under QEMU, the values are (sometimes) 0xff,0xff after a reset, which caused init to fail (sometimes).

Because KOS only supports *normal* ata, just remove the non-working detection code and let it error out in one of the other places when nothing's connected.