Improve detection of global object memory leaks
The improvement is actually to do less, and with this change, the memfd system-test actually shows up as a leak now! (Yay: the test cases no longer run to completion)

But seriously, this memfd-leak is something that I already knew about, and is actually the reason I embarked on the idea of improving the memory leak detector to the point where it would detect it as a leak. Now that that's happening, the sub-system can be re-designed to fix the leak.