Add a linux-compatible /proc/modules
Though some of the info linux puts into that file isn't actually relevant on KOS, such as the refcnt field.
On KOS, the kernel does all of the module loading/unloading, which includes keeping track of when/how drivers are supposed to be unloaded.

This file is primarily being provided to allow for linux-compatible lsmod(1) programs to run under KOS.