Fix a bug related to writing to unloaded file areas
There was some missing handling for this special case, which instead assumed that the written-to area wasn't allocated yet, and thus filled it with zeroes, leading to file corruption in the form of certain areas of files suddenly turning to zeroes.

That is now fixed, and a system-test that could reproduce the problem (but no longer does) has also been added.