New program /bin/cc
Similar to /bin/leaks, this is an x86-specific, minimal-overhead program that simply invokes the system_cc() kernel function.
Since said function is all about clearing system caches, having a program which itself has a minimal impact on caches is quite useful as any cache effects by the program making the request are minimal this way.