Improve `nomall' commandline option
The option now fully disables the trace-malloc system, re-configuring the kernel to behave the same as it would when compiled with `CONFIG_NO_TRACE_MALLOC'.

Also: the implementation for this no longer resides in arch-specific code, but instead uses a new, optional arch-specific API <asm/redirect.h>, making the kernel a little bit more portable.