Implement PE dependency loading
Well here's a problem: that super-simple test-exe that's generated by `make_utility.sh i386 testexe` has a dependency on KERNEL32.dll, which is something it's never going to find under KOS (at least not for quite a while). So in other words: it looks like I'll have to find a better method of generating test PE files that I can then try to run under KOS.
Though I should mention that the PE loader correctly detects the dependency and exits with an error message telling you exactly that.