Fix kernel crash on iret with EFLAGS.NT=1
Problem with this one was that user-space programs are able to use popf to set EFLAGS.NT=1, which would then lead to a kernel crash later down the line.
Resolve this problem by detecting the condition in the relevant kernel exception handler, and fixing it by clearing EFLAGS.NT before trying again.