Add user-space support for nopf
Also add integrity checks for the TLS segment in exception handling. -- If a user-space program faults while having a faulty ELF TLS context, rather than have the exception handler end up causing a stack overflow, the handler will detect the corrupted TLS context and immediatly trigger a coredump.

Note that user-space is currently lacking a header to peroperly expose nopf functionality, and that this sort of functionality as a whole is entirely architecture-specific (meaning there can't be a portable implementation). However, what's currently missing will be added sooner or later.