Add support for loading 32-bit PE files under x86_64
Not much was really missing here, except for the kernel's PE loader to detect compatibility-mode files, and set-up the execution-environment accordingly. -- After all: we've already been supporting execution of 32-bit ELF files under a 64-bit kernel since the beginning, so said environment was already there.

But note that this commit also fixes an bug in `libdl-pe.so' that just happened to not get triggered under 32-bit, where loading additional dependencies caused a dereft from an uninitialized `%segtls:0'. -- That's been fixed now!