Simplify & optimize user-space pointer validation
By working under the assumption that user-space structures aren't accessed in access-deltas > PAGESIZE (which already was an assumption at this point), we can simplify pointer validation by only checking the base address.
This way, the actual # of validate_* functions can be reduced drastically, and those that remain can be simplified by only needing to check the base address (as opposed to testing for overflow and similar).

Note though that these validation functions will continue to take access-size arguments, in order to keep it possible to replace these macros with something more interesting in the future.

Also: to ensure that the kernel correctly catches invalid user-spaces pointers near the user/kernel boundary, expand the sysexcept system-test to make a bunch of system calls that pass such invalid pointers.