Fix `TIOCSCTTY' not setting session->tty
Ooops. This caused open("/dev/tty") to fail with E_NO_CTTY. It's a wonder that everything worked despite this oversight, but I guess that as far as the terminal was concerned, it was the controlling terminal of the initial session. -- Only that said session didn't remember that it had assigned itself a controlling terminal.