Add handling for gnulib autoconf checkers
Wow! That was a lot of work, but these should cover _any_ gnulib system feature test that uses AC_RUN_IFELSE(), and thus wouldn't work in cross-compilation, causing it to have to make assumptions. With this, we can explicitly tell gnulib what KOS is implementing, without it trying to override our system headers all of the time.