Reduce the size of the kernel-internal `sigset_t' to 8 bytes
... from what used to be 128 bytes.

It now has the minimal-possible size of 64 bits (8 bytes), which is able to house all currently defined signals (s.a. NSIG).

User-space is unaffected by this, as all system interfaces have been updated/altered to accept/return variable-sized signal masks, meaning that new signals can now be added to the kernel without breaking anything in user-space!

Note that this is a significant change to the kernel ABI, so to build this commit, you probably have to do a forced full-rebuild.