Cleanup QEMU configuration
Use a file hidden from the git to facilitate tinkering and hacking. This primary fixes this script tying to access a file "$PROJDIR/../disk.img" that's actually located outside of teh .git root directory. (I was using this when programming the tarfs driver by putting a tar file under that name)