Improve integration of new handman system
This commit also starts making use of the new 2-step handle installation system, and finally fixes the potential scope leakable and race condition of creation a new file, only to have the open(2) call fail at the very last second because it was unable to allocate a new handle.

With this, the handle is allocated first, such that once the file has been created, nothing can fail anymore, and the system call is guarantied not to fail.