Add support for positional printf (e.g. "%1$s")
This can also be done in glibc, and with this commit it also becomes possible under KOS.
Note that we don't impose limits on many positional arguments are used, however we _do_ make use of alloca(), so if you go too over-board with them, you'll end up with a stack overflow.
Note that for now positional arguments are enabled for in libc, as well as in the kernel (though not in libdl). However, nothing in kernel-space actually uses them (for now), and the only user-space consumer is a newly added system-test that's there to ensure that positional arguments work as intended (and as documented by glibc's `man 3 printf')