Optimize pp expressions in libc auto sources
Optimization is done by making the assumption that !defined(__KERNEL__) implies that all of the `__CRT_HAVE_*' macros from `#ifndef __KERNEL__'-blocks in libc's internal crt-features file are defined.

This doesn't actually change anything semantically, but allows for a whole bunch of code that was essentially `#if 0' to be deleted during the code generation phase.