Use st_btime and st_ctime for their intended purposes
And with this, all 4 time fields exist in--, and are used for their proper prupose in the KOS kernel!

Also note that there is no system call to explicitly set the st_ctime field, as the AT_CHANGE_CTIME flag has been changed to AT_CHANGED_BTIME, such that it changes a file's birth time (as has always been its intended purpose; it was just badly named since I always thought that st_ctime was the "creation" time field, when it fact its the "changed" field)
Welp: I've learned since, and now KOS is more compliant as a result!