Add dwarf-5 support to <libdebuginfo/debug_line>
With this, we can (currently only via the super-slow try-hard path technically only meant for when debug-info is corrupted) once again do basic addr2line. But note that porting to dwarf-5 isn't completed yet (hence me having yet to change the `-gdwarf-4` to `-g` in `kos/.sources`)

Also note that all changes made should be backwards compatible with dwarf-4 (trying to debug some of the old dwarf-4 programs shows no changes, so it appears that I havn't broken any of them), and that this backwards-compatibility with be retained indefinitely!