Unify system configurations
Introduce a new "options" database that collects and maintains all of the available system configuration macros. - The idea is that this database can eventually be used to easily create custom system configurations (beyond the 4 default configs OD, nOD, OnD and nOnD), which then define macros from the options database in a user-defined manner.