Fix MIN/MAX macros defined in <sys/param.h>
Some code assumes that these can be used in constant expressions. As such, we have to define them in a way that complies with this.

Additionally, add support for the long-since removed g++ extensions for implementing min/max constant expressions that only evaluate their arguments once. (Because fun fact: intellisense actually understands this syntax, meaning we can use it to speed up IDE parsing)