Make regex fastmap direct dispatch easier
Done by repeating top-level prefix opcodes for each alternation of a top-level group. This way, the fastmap generator will not be hindered by opcodes with side-effects being present prior to the first top-level set of alternations. (iow: `^(foo|bar)$` is compiled as `(^foo|^bar)$`, so that the fastmap can directly dispatch "f" and "b" to the relevant alternation, without the need fo an on-failure stack entry)