Upgrade to latest version of deemon
Hopefully, this will make it easier to build KOS on a unix machine (since this deemon upgrade now includes the re-written ipc module with focus on compatibility with unix).

Note that building KOS under windows still works just as it did before (and I know for a fact that building KOS on a unix machine probably had major problems before (which it still might have; I'll have to test that)), so even if there are still problems when you try to build KOS under unix, it'll at least (hopefully) be less/others than those you faced before.

Also note that this change in deemon requires a couple of