Enforce proper locking for functions like `printf()`
Acquire and maintain the same lock for the duration of the call, rather than acquiring a lock every time some part of the string is printed. -- This way, it become impossible for 2 threads both printing to the same time to have their output become intermangled.

This also speeds up formatted stdio print functions by quite a bit.