Add support for auto-mmap of files during `read(2)`
With this, the kernel will automatically serve very large calls to `read(2)` by creating a misaligned mfile wrapper which is then mmap'd in such a way that it replaces the pre-exiting buffer mapping.

With this, the traditional way of using `open()+malloc()+read()` to load a file into memory can now be just as fast and efficient as doing an `open()+mmap()`. The threshold that determines if the kernel does this for some given read-request is runtime configurable via `/proc/kos/mm/part-automap-threshold`, and the feature in its entirety can be hard-disabled by building the kernel with `-DCONFIG_NO_AUTO_MMAP`

As a test & demonstration of this feature, you can run `playgroup lread`