Better `E_SEGFAULT` handling while holding mman locks
With this, the #PF handler now tries to detect situations where it's not able to lock the mman, which are then handled by checking if the faulting PC is in a kernel `NOTHROW` function. If it it, that means that the memory access *can* be NCX-related, meaning that it *has* to be a bug. This way, a dead-lock is prevented, which *should* then allow the kernel to at least pring a traceback.