Automatically unshare misaligned files during writes
This fixes the problem documented by the previous commit. And with this, I can no longer think of a scenario where mmapread can be distinguished from a regular read (other than mincore introspection, and obviously excluding behavior related to bugs I've yet to find). There are still 2 TODOs remaining, one of which is actually related to semantics (meaning that until it's resolved, there *is* a small race condition that *could* result in mmapread data not being unshared when it should be, and another that's needed to for memory optimization).