Hopefully fix the state-corruption bug in `tsc.S`
When forcing this code-path to be used (by uncommenting the `je     .Lapic_ipi_no_pending` in L#197), the absence of this line was able to consistently trigger the fault in `tsc.S` (because the CPU state wasn't written to the origin task), but by **actually** saving the state, that no longer happens.

I was technically unable to force the bug to happen *without* having `L#197` commented out, but assuming the kernel *did* sometimes want to switch tasks as the result of an IPI, this most definitely *would* have caused it to fault in that case.

So I'm just going to assume that I **did** find the bug, and move on.